Split kvstore file to: fwtree, estore & kvstore
